Move `RemoteControl.CodeIntelligence.Ast` into `Ast`
This module provided only a relatively thin wrapper over `Aliases.at/2`. It has
been moved into `Lexical.Ast`.
